--- src/core/routing/useCanGoBack.ts
@@ -1,9 +1,0 @@
-import { useLocation } from 'react-router-dom';
-
-const useCanGoBack = () => {
-  const location = useLocation();
-  const canGoBack = location.key !== 'default';
-  return canGoBack;
-};
-
-export default useCanGoBack;
