+++ src/crd/i18n/layout/layout.bg.json
@@ -1,5 +1,6 @@
 {
   "header": {
+    "skipToContent": "Към съдържанието",
     "menu": "Меню",
     "home": "Начало",
     "search": "Търсене",
@@ -21,17 +22,7 @@
     "changeLanguage": "Смяна на езика",
     "getHelp": "Помощ",
     "showGrid": "Покажи мрежа",
-    "hideGrid": "Скрий мрежа",
-    "designVersion": {
-      "label": "Нов облик (старият дизайн е временно достъпен)"
-    },
-    "designVersionUpgrade": {
-      "title": "Свежият нов Alkemio е тук",
-      "body": "Пускаме нов дизайн, който е по-бърз, по-ясен и готов за това, което предстои. Каним те да преминеш сега.\n\nАко предпочиташ стария вид, той ще остане достъпен от менюто на твоя аватар за ограничено време - а ако нещо ти липсва, моля свържи се с поддръжката, за да можем да го пренесем, преди старият дизайн да бъде премахнат.",
-      "confirm": "Към новия дизайн",
-      "dismiss": "Може би по-късно",
-      "closeLabel": "Затвори"
-    }
+    "hideGrid": "Скрий мрежа"
   },
   "footer": {
     "copyright": "© 2026 Alkemio B.V.",
