+++ src/crd/i18n/layout/layout.de.json
@@ -1,5 +1,6 @@
 {
   "header": {
+    "skipToContent": "Zum Inhalt springen",
     "menu": "Menü",
     "home": "Startseite",
     "search": "Suchen",
@@ -21,17 +22,7 @@
     "changeLanguage": "Sprache ändern",
     "getHelp": "Hilfe erhalten",
     "showGrid": "Raster anzeigen",
-    "hideGrid": "Raster ausblenden",
-    "designVersion": {
-      "label": "Neuer Look (altes Design vorübergehend verfügbar)"
-    },
-    "designVersionUpgrade": {
-      "title": "Das neue Alkemio ist da",
-      "body": "Wir führen ein neues Design ein - schneller, klarer und bereit für das, was kommt. Du bist eingeladen, jetzt zu wechseln.\n\nWenn du das alte Aussehen bevorzugst, bleibt es für eine begrenzte Zeit über das Menü deines Avatars verfügbar - und falls etwas fehlt, melde dich bitte beim Support, damit wir es übertragen, bevor das alte Design eingestellt wird.",
-      "confirm": "Zum neuen Design",
-      "dismiss": "Vielleicht später",
-      "closeLabel": "Schließen"
-    }
+    "hideGrid": "Raster ausblenden"
   },
   "footer": {
     "copyright": "© 2026 Alkemio B.V.",
