+++ src/crd/i18n/layout/layout.en.json
@@ -1,5 +1,6 @@
 {
   "header": {
+    "skipToContent": "Skip to content",
     "menu": "Menu",
     "home": "Home",
     "search": "Search",
@@ -21,17 +22,7 @@
     "changeLanguage": "Change Language",
     "getHelp": "Get Help",
     "showGrid": "Show Grid",
-    "hideGrid": "Hide Grid",
-    "designVersion": {
-      "label": "New look (old design available for a limited time)"
-    },
-    "designVersionUpgrade": {
-      "title": "A fresh new Alkemio is here",
-      "body": "We're rolling out a new design that's faster, clearer, and built for what's next. You're invited to switch now.\n\nIf you prefer the old look, it'll stay available from your avatar menu for a limited time - and if something is missing, please let support know so we can carry it over before the old design retires.",
-      "confirm": "Take me to the new design",
-      "dismiss": "Maybe later",
-      "closeLabel": "Close"
-    }
+    "hideGrid": "Hide Grid"
   },
   "footer": {
     "copyright": "© 2026 Alkemio B.V.",
