+++ src/crd/i18n/layout/layout.es.json
@@ -1,5 +1,6 @@
 {
   "header": {
+    "skipToContent": "Saltar al contenido",
     "menu": "Menú",
     "home": "Inicio",
     "search": "Buscar",
@@ -21,17 +22,7 @@
     "changeLanguage": "Cambiar idioma",
     "getHelp": "Obtener ayuda",
     "showGrid": "Mostrar cuadrícula",
-    "hideGrid": "Ocultar cuadrícula",
-    "designVersion": {
-      "label": "Nuevo aspecto (diseño antiguo disponible temporalmente)"
-    },
-    "designVersionUpgrade": {
-      "title": "El nuevo Alkemio ya está aquí",
-      "body": "Hemos diseñado una nueva versión más rápida, más clara y preparada para lo que viene. Te invitamos a cambiar ahora.\n\nSi prefieres el aspecto antiguo, seguirá disponible desde el menú de tu avatar durante un tiempo limitado - y si echas algo de menos, avisa al equipo de soporte para que podamos trasladarlo antes de retirar el diseño antiguo.",
-      "confirm": "Llévame al nuevo diseño",
-      "dismiss": "Quizás más tarde",
-      "closeLabel": "Cerrar"
-    }
+    "hideGrid": "Ocultar cuadrícula"
   },
   "footer": {
     "copyright": "© 2026 Alkemio B.V.",
