+++ src/crd/i18n/layout/layout.nl.json
@@ -1,5 +1,6 @@
 {
   "header": {
+    "skipToContent": "Ga naar inhoud",
     "menu": "Menu",
     "home": "Home",
     "search": "Zoeken",
@@ -21,17 +22,7 @@
     "changeLanguage": "Taal wijzigen",
     "getHelp": "Hulp krijgen",
     "showGrid": "Raster tonen",
-    "hideGrid": "Raster verbergen",
-    "designVersion": {
-      "label": "Nieuwe look (oud ontwerp tijdelijk beschikbaar)"
-    },
-    "designVersionUpgrade": {
-      "title": "Het vernieuwde Alkemio is er",
-      "body": "We rollen een nieuw ontwerp uit dat sneller, helderder en klaar voor de toekomst is. Je bent uitgenodigd om nu over te stappen.\n\nWil je liever het oude uiterlijk? Dat blijft een korte tijd beschikbaar via het menu onder je avatar - en mis je iets, laat het support weten zodat we het kunnen overzetten voordat het oude ontwerp wordt uitgefaseerd.",
-      "confirm": "Breng me naar het nieuwe ontwerp",
-      "dismiss": "Misschien later",
-      "closeLabel": "Sluiten"
-    }
+    "hideGrid": "Raster verbergen"
   },
   "footer": {
     "copyright": "© 2026 Alkemio B.V.",
