+++ src/domain/collaboration/whiteboard/containers/useWhiteboardActions.ts
@@ -5,7 +5,7 @@ import {
   WhiteboardDetailsFragmentDoc,
 } from '@/core/apollo/generated/apollo-hooks';
 import type { CreateWhiteboardInput } from '@/core/apollo/generated/graphql-schema';
-import { evictFromCache } from '@/core/apollo/utils/removeFromCache';
+import { evictFromCache } from '@/core/apollo/utils/evictFromCache';
 import type { Identifiable } from '@/core/utils/Identifiable';
 import useUploadWhiteboardVisuals from '../WhiteboardVisuals/useUploadWhiteboardVisuals';
 import type { WhiteboardPreviewImage } from '../WhiteboardVisuals/WhiteboardPreviewImagesModels';
